+++ xcloud-iam-core/src/main/java/com/wl4g/iam/web/ContactController.java
@@ -54,9 +54,9 @@ public class ContactController extends BaseController {
 		log.info("into ProjectController.save prarms::" + "contact = {} ", contact);
 		RespBase<Object> resp = RespBase.create();
 		Assert.notNull(contact, "contact is null");
-//		Assert.hasText(contact.getName(), "name is null");
-//		Assert.hasText(contact.getEmail(), "email is null");
-//		Assert.notEmpty(contact.getGroups(), "contactGroup is null");
+		// Assert.hasText(contact.getName(), "name is null");
+		// Assert.hasText(contact.getEmail(), "email is null");
+		// Assert.notEmpty(contact.getGroups(), "contactGroup is null");
 
 		contactService.save(contact);
 		return resp;
@@ -64,7 +64,7 @@ public class ContactController extends BaseController {
 
 	@RequestMapping(value = "/detail")
 	@RequiresPermissions(value = { "iam:contact" })
-	public RespBase<?> detail(Integer id) {
+	public RespBase<?> detail(Long id) {
 		log.info("into ContactController.detail prarms::" + "id = {} ", id);
 		RespBase<Object> resp = RespBase.create();
 		Contact contact = contactService.detail(id);
@@ -74,7 +74,7 @@ public class ContactController extends BaseController {
 
 	@RequestMapping(value = "/del")
 	@RequiresPermissions(value = { "iam:contact" })
-	public RespBase<?> del(Integer id) {
+	public RespBase<?> del(Long id) {
 		log.info("into ContactController.del prarms::" + "id = {} ", id);
 		RespBase<Object> resp = RespBase.create();
 		contactService.del(id);
